requirements.txt
asgiref==3.11.1
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
sqlparse==0.5.5
